src/sql_agent/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
python-dotenv>=1.0.1

sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pandas>=2.2.2

streamlit>=1.36.0
requests>=2.32.3

langchain>=0.2.10
langgraph>=0.2.15
langchain-openai>=0.1.20
langchain-community>=0.2.10

openai>=1.40.0

langfuse>=2.35.0
comet-ml>=3.43.0

